refactor cache handler
